Recreate truncates if file exists
Truncate the file to 0 bytes opened with uproot.recreate before writing contents. This behavior agrees with ROOT Recreate.
The previous code lines seemed to intend to do this, however assumingly due to an erroneous 'not', this was not done.
If the file does not exist, an empty file is created. This is the same behavior as before.