All of the B023 warnings are okay in this file.
I don't know why the v4 branch is complaining, whereas the v5 branch is not, and these should (ideally) be handled more centrally than tagging each use—that was probably done on the v5 branch. Nevertheless, we don't plan to develop this branch much more, just bug-fixes, so obfuscating it with "`# noqa: B023`" is not a problem here—just don't do it in the `main` (v5) branch!