fix: Only dispatch requests to Source if we actually will use them (#1436)
Only dispatch requests to Source if we actually will use them

The notifications queue in _ranges_or_baskets_to_arrays cannot be aware
of chunk requests in flight because the queue only receives the returned
chunks. If we have the necessary array in cache, then we exit this
function and fire-and-forget chunk requests causing extra network
traffic, and more to the point, potential race conditions between
receiving chunks and closing the file on teardown.

Fixes https://github.com/scikit-hep/coffea/issues/1331